Fix failure because of race conditions for update-resolv-conf.service

+++ pkg/webhook/controlplane/ensurer.go
@@ -239,6 +239,8 @@ StartLimitIntervalSec=0
 [Service]
 Type=oneshot
 ExecStart=/opt/bin/update-resolv-conf.sh
+RestartForceExitStatus=10
+RestartSec=15
 `
 	)
 
@@ -283,6 +285,10 @@ is_systemd_resolved_system()
       return 1
     fi
 }
+
+if grep -q "# No DNS servers known." /run/systemd/resolve/resolv.conf; then
+  exit 10
+fi
 
 rm -f "$tmp"
 if is_systemd_resolved_system; then
